Science update and Exploration
Overall of the science department.
Moves the RD's office to up above, adds in a break room, science bathroom, chemistry lab. Moves the RD servers. Adds in a workshop for circuts. Expands robotics a bit, expands the storage room.

Adds in exploration.
- Pathfinder office
- Gearing up room
- Meeting room/break room
- Hanger
- Gateway

Only downside is, the shuttle and the gateway do not work at the moment. But we'll have it at least so someone can go in and fix them easier.