A variety of fixes
Pixal shifts the captains consoles a bit more
Moves the bridge fire-axe to a better location
Adds some tables/vending machines in the command hallway
Adds some tables outside of the command hallway
Adds more rechargers by the bridge, and in the captains office
Fixes the cell's in Security, so each has it's proper tag on the flashers, lockers, and buttons.
Moves the trash pile that was blocking the way into maint.
Moves the air alarm in the HOPS office so it's more acceessable
Adds armor to the blueshield's armor locker. Not sure what else to add, so will leave it up to others.
removed the unneeded ceiling tiles on deck two by cargo
Added a light switch to the chapel